js/xpconnect/tests/mochitest/test_bug862380.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=862380
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 862380</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 862380 **/
  SimpleTest.waitForExplicitFinish();
  function go() {
    checkNotEnumerable($('ifr').contentWindow);
    checkNotEnumerable($('ifr').contentWindow.location);
    SimpleTest.finish();
  }

  function checkNotEnumerable(obj) {
    try {
      is(Object.getOwnPropertyNames(obj).length, 0, "Object.getOwnPropertyNames gives empty array");
      is(Object.keys(obj).length, 0, "Object.keys gives empty array");
      for (var i in obj)
        ok(false, "Enumerated something: " + i);
    } catch (e) {
      ok(false, "threw: " + e);
    }
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=862380">Mozilla Bug 862380</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<iframe id="ifr" onload="go();" src="http://example.org/tests/js/xpconnect/tests/mochitest/file_empty.html"></iframe>
<pre id="test">
</pre>
</body>
</html>
